I consolidated the drop-down menu using javascript such that it's reusable on all pages via the component script.
I also noticed I named a file on the cv index as index.css, it worked fine but for standard, I renamed it style.css and corrected the file path on the html file as well.

What else? Fewer lines of code and easier to maintain. 
Like I said, I'll consolidate more...